Changed all unsigned int to int.
Also made sure that UINT_MAX is INT_MAX.

+++ lib/test_ercs.c
@@ -29,9 +29,9 @@
 
 
 static ercs_t *
-alloc_ercs(double L, unsigned int num_points, unsigned num_loci) 
+alloc_ercs(double L, int num_points, int num_loci) 
 {
-    unsigned int j;
+    int j;
     ercs_t *sim= xcalloc(1, sizeof(ercs_t));
     sim->num_loci = num_loci;
     sim->num_parents = 2;
@@ -42,7 +42,7 @@ alloc_ercs(double L, unsigned int num_points, unsigned num_loci)
     sim->torus_diameter = L;
     sim->sample_size = num_points;
     sim->max_time = DBL_MAX;
-    sim->max_kdtree_insertions = UINT_MAX;
+    sim->max_kdtree_insertions = INT_MAX;
     sim->sample = xmalloc(2 * num_points * sizeof(double));
     sim->event_classes = xmalloc(sizeof(event_class_t));
     sim->recombination_probabilities = xmalloc(num_loci * sizeof(double));
@@ -55,7 +55,6 @@ alloc_ercs(double L, unsigned int num_points, unsigned num_loci)
     }
     sim->recombination_probabilities[num_loci - 1] = 0.0;
     alloc_disc_event_class(sim->event_classes, 1.0, L / 10.0, 0.95); 
-
     return sim;
 }
 
@@ -78,8 +77,8 @@ static int
 compare_histories(ercs_t *sim1, ercs_t *sim2)
 {
     int ret = 0;
-    unsigned int l, j;
-    unsigned int n = sim1->sample_size; 
+    int l, j;
+    int n = sim1->sample_size; 
     for (l = 0; l < sim1->num_loci; l++) {
         for (j = 0; j < 2 * n; j++) {
             if (sim1->pi[l][j] != sim2->pi[l][j]) {
@@ -98,18 +97,18 @@ out:
 }
 
 static int 
-test_ercs_size(double L, unsigned int num_points, unsigned int m)
+test_ercs_size(double L, int num_points, int m)
 {
     int ret = 0;
     int not_done;
-    unsigned int num_events;
+    int num_events;
     ercs_t *sim1 = alloc_ercs(L, num_points, m);
     ercs_t *sim2 = alloc_ercs(L, num_points, m);
     ret = ercs_initialise(sim1); 
     ERCS_ERROR_CHECK(ret, out);
     ret = ercs_initialise(sim2); 
     ERCS_ERROR_CHECK(ret, out);
-    ret = ercs_simulate(sim1, UINT_MAX); 
+    ret = ercs_simulate(sim1, INT_MAX); 
     ERCS_ERROR_CHECK(ret, out);
     not_done = 1;
     num_events = 0;
@@ -120,7 +119,6 @@ test_ercs_size(double L, unsigned int num_points, unsigned int m)
         num_events++;
     }
     ret = compare_histories(sim1, sim2);
-    //ercs_print_state(sim1);
     free_ercs(sim1);
     free_ercs(sim2);
 out:
@@ -134,12 +132,12 @@ static int
 test_ercs(void)
 {
     int ret = 0;
-    unsigned int j, k, n, m;
+    size_t j, k, n, m;
     double torus_sizes[] = {1.9, 0.2, 50.3, 123.22, 1111.1};
-    unsigned int num_points[] = {2, 11, 50, 212, 333, 515};
+    int num_points[] = {2, 11, 50, 212, 333, 515};
     printf("\trunning ercs tests");
     for (j = 0; j < sizeof(torus_sizes) / sizeof(double); j++) {
-        for (k = 0; k < sizeof(num_points) / sizeof(unsigned int); k++) {
+        for (k = 0; k < sizeof(num_points) / sizeof(int); k++) {
             for (m = 1; m < 4; m++) {
                 printf(".");
                 fflush(stdout);
